Various Changes and fixes
- New Features -
CustomUI Namespace - I created a customUI namespace the has the ability to hold any custom UI we want to create or use.

PropertiesPannel class - I created a properties pannel class inside of the CustomUI namespace that handles what the UI for how the user interacts with gameobjects.

- Changes -
Moved Functions - I move the majority of the function in Scene1 into Scene to save space and clean up the code.