-- Additions --
- SaveFile Class
This class handles holds all of the needed containers for the save files and includes a struct named "ElementInfo" that holds all of the needed information for a XMLElement. In addition a enum class named "FileType" was added. This control's how the save file will be inturpreted when being saved and loaded and includes different file types and directories for each type.

- SaveManager Class
This class utilizes static variables and functions that store all of the savefiles and handles the insertion of savefiles and a SaveQueue system that allows us to only save files that have been edited.

- SaveUtility Class [IN DEVELOPMENT]
This singleton class allows gives us the ability to edit save files and add/edit information that is contained in them. It also allows us to Create/Delete/Overwrite save files.

-- Changes --

- Globals Class
 Added Project path to Globals class to allows ease of access.