-- Changes -- GameObject Class: The gameobjects name was changed from const char * to std::string. A new bool was also added to allow objects to be instantated at runtime.
SceneGraph Class: Added a unordered map the contans instantiable objects and made fixes to how objects are instantiated.

SaveFile Class: Added a helper function to check if a save file contains an element.

SaveFile Class: Added function that transfers a save from the SaveFiles map to the SaveQueue. And a helper function that returns if ther is a save file of type currently saved.

SaveUtility Class: Changed the logic for some functions to allow for a more dynamic saving experience.

Scene Class: Changed the function that handles objects being spawned In, Minor change to UI layout for spawnable objects

ALL GameObjects: All changed to use std::string instead of const char* for name.