------ Changes ------
SceneGraph Class: Added function that checks if a object with a name is already in the scene for use in the AddGameObject function.

MeshRenderer Class: Added GetModel function to allow for future use in the properties panel.

CustomUI namespace: Added the ability to drag models from the content browser to a mesh renderer slot in the properties panel and change the render flags. You can also now change the active render view mode. (Albedo, Pos, etc)

UIStatics Class: Added functions to Draw Vec4's and TextureSlots for use in the properties panel.