Quaternion Rotations & Save File Deletion
------ Changes ------

RigidBody Class: Work on converting rotational velocity from pure vector to changing the transform quaternion has started.

Transform Class: The rotation vector has now been changed to a quaternion and the needed changes have been added to properly represent rotations via quaternions.

SaveFile Class: Some helper functions and variables have been added to allow for the old save object file to be deleted if the name is changed in the hierarchy.

SaveManager Class: Function added to allow a save name to be changed from the map that contains them.

CustomUI Namespace: The dockspace now contains a vector of UIInterfaces that are updated and rendered.

MMath Class: Added some helper functions to allow a matrix3 to be converted to a matrix4 for getting the rotational matrix.

Quaternion Class: Some functions were changed to fix some of the syntax in equations.