------ Additions ------ Task Class: Holds a std::function pointers to whatever function this task is currently loaded with and executes it once RunTask is called.
Thread Class: Handles all of the functionalities need for a thread and is also used as a wrapper for std::thread so we can have for usability within each thread.

ThreadHandler Class: Creates and manages all of the threads that are running In the thread pool and assignents tasks to them.

------ Changes ------

Minor fixes to some floating point warnings in mutlple different classes.