------ Changes ------
Task Class: Added ETaskType enum so that the thread handler will know what thread that task needs to run on if specified.

Thread Class: Added EThreadType so the thread will know what type of tasks it can run. (Needed because any OpenGL calls can only be run on a single thread at a time)

ThreadHandler Class: Added a mutex and conditional varible to make sure that all our functions are thread safe.

Network Managers: Deleted the Pre-Created std::thread and connected our thread pool to work with it.

ModelManager Class: Turned it into a Singleton because in order to add tasks from this class we either need a pointer to the class/object or a refrence.