----- Changes -----
Task Class: Made some minor changes to ETaskType and how it will work in the system.

Thread Class: Made sure that if a thread was flagged for rendering operations that the opengl context would switch to the new thread.

ThreadHandler Class: Added functionality to allow thread types and task types to be synced up so that a rendering task can't get called on a general thread. (Still needs work)

Window Class: Added another OpenGL context for the seperate thread and added new function to initalize that context and clear the old one.