----- Changes -----
CoreEngine Class: All rendering initalization has been switched to a new rendering thread and all rendering is now done on that thread.

TextureManager Class: Switched to a singleton for future use with the thread pool.

Thread Class: Removed some un-needed code as we will be doing rendering on one thread and not switching between them.

ThreadHandler Class: Added the ability to add a task and pass in flags to make the function wait for that task to finish before finishing.