Pathfinding System
----- Additions -----

Pathing Algorithms Class: Contains a namespace with the base PathAlgorithm classes and then contains both the BreadthFirst & AStar algorithms as PathAlgorithms.

AIDirector Class: Is gonna be used more in the future but for not it just allows us to find a path with the passed in algorithm via an enum.

----- Changes -----
- Moved some of the steering classes into another folder so I can keep it organized.