------ Additions ------ Poly Class: Just used so that we can store a polygons list of verticies and some other helper functions (Will be used more once we get to NavMesh.
NavigationMesh Class: Will handle the proccess of converting given verticies into a navigation mesh that can be converted to a graph for pathfinding operations.

------ Changes ------

DivisionAlgorithms Namespace: Stores and holds all of the algorithms that we can use to create a visability graph.

AIDirector Class: Is used to manage all of the pathfinding related operations and sends map data to the given algorithms.

CustomUI Namespace: Fixed bug that didn't remove an object from the scene savefile when deleted from scene.

VMath/Vector Class: Added some functions to support 2D raycast for creating the visibilty graph.

Ray Class: Added a function that checks if the ray is passing through a given line(Two Vec2's)