Collision Detection Changes
------ Additions ------

BoundingSphere Class: Added class that will create a bounding sphere around an object with radius and position.

------ Changes ------

Rigidbody3D: Added a ConstructCollider function that will initialize the collider object depending on the parameters.

GameObject Class: Made PostInit call the rigidbody3d AddCollisionFunction because we now allow the gameobject to tell the rigidbody what collider it has.

BoundingBox Class: Made some changes to how the World_Verticies vector worked and where it was contained.