Scene Cleanup/ Load Saves Fix
------ Changes ------

Globals File: UIStatics & Globals have been merged into one namespace called "Globals" this should be where we store any global info from now on.

 CoreEngine Class: Moved gizmo switching controls into the handle events function, Save/Load Scene has also been moved in.

LoadUtility Class: Function was created that will fix part of the issue where having where when adding new data to the LoadObjectFunction we would need to go in and edit the file for it to load properly.