------ Changes ------
We now have some reflection!

A C# script was created to look through some of our common gameobject directories, take their class name and file path. It then uses that information to write code in the LoadUtility.cpp AddObjectToMap function so that we no longer need to add anything manually.
It also creates a new include file and stores the includes to all of the needed gameobjects :)